Move functions to class methods
