+++ project/rsm.py
@@ -22,3 +22,16 @@ class RSM(NamedTuple):
 
     start_symbol: Variable
     boxes: Dict[Variable, DeterministicFiniteAutomaton]
+
+    def minimize(self) -> "RSM":
+        """Minimizes RSM by minimizing internal automatons
+
+        Returns
+        -------
+        minimized_rsm : RSM
+            Minimized recursive state machine
+        """
+        return RSM(
+            start_symbol=self.start_symbol,
+            boxes={v: a.minimize() for v, a in self.boxes.items()},
+        )
